refactor: move attack controls to a live component
Abstracting the attack controls out into a separate live_component so that the
state can be managed outside of the quest view directly.

Hides the attack controls if the user does not have an adventurer.

Uses semantic HTML buttons for the attack selectors instead of divs with button
styling.
